bug(timestamps): test exercising issue with timeless leakage.
Timeless is implemented using a counter incremented twice if both created_at
and updated_at are included (as they are by simply using Mongoid::Timestamps,
but the test for timelessness only lowers counter by 1. So one of the timeless-
nesses leaks out for each call to timeless.

Worse is that timelessness on created_at breaks the whole idea of being able to
rely on timestamped records always having a created_at. What possible utility is
this?